apply mentioned comments
