+++ src/hw-01/script1.js
@@ -1,3 +1,6 @@
+/** Выводит в консоль произведение и сумму чисел, заданных константами.
+ * Пример.
+ */
 export default function printToConsole() {
   const a = 10;
   const b = 2;
