+++ src/hw-01/script2.js
@@ -1,0 +1,7 @@
+/** выведет в консоль суммарное количество символов в строках, заданных входными параметрами.
+ * @param {string} s1
+ * @param {string} s2
+ */
+export default function printConcatLength(s1, s2) {
+  console.log(s1.length + s2.length);
+}
