+++ src/hw-03/script3.js
@@ -1,4 +1,5 @@
-// 1.Вывести в консоль сумму всех целых чисел от 50 до 100.
+/** Выводит в консоль сумму всех целых чисел от 50 до 100.
+ */
 export function printSum50to100() {
   let result = 0;
   for (let i = 50; i <= 100; i += 1) {
@@ -7,12 +8,8 @@ export function printSum50to100() {
   console.log(result);
 }
 
-/* 2.Вывести в консоль таблицу умножения на 7.
-7 x 1 = 7
-7 x 2 = 14
-…
-7 x 9 = 63
-*/
+/** Выводит в консоль таблицу умножения на 7.
+ */
 export function printTableMultBy7() {
   const seven = 7;
   for (let i = 1; i < 10; i += 1) {
@@ -21,12 +18,10 @@ export function printTableMultBy7() {
   }
 }
 
-/* *Запросить у пользователя ввод числа N. Вывести в
-консоль среднее арифметическое всех нечётных
-чисел от 1 до N.
-*/
+/** Выводит в консоль среднее арифметическое всех нечётных чисел от 1 до N, запрощенное у пользователя.
+ */
 export function calcAvgOfOdds() {
-  const n = +prompt("Введите число: ", 0);
+  const n = Number(prompt("Введите число: ", 0));
 
   let res = 0;
   let j = 0;
