+++ src/hw-07/index7.js
@@ -1,0 +1,5 @@
+import { addForm, addContent } from "./script7.js";
+
+addContent(document.querySelector("#content"));
+
+addForm(document.querySelector("#form"));
