+++ src/hw-09/index9.js
@@ -1,0 +1,3 @@
+import { printRootsOfQEquation } from "./script9.js";
+
+printRootsOfQEquation();
