+++ src/hw-09/script9.test.js
@@ -15,11 +15,6 @@ import {
 уравнения с коэффициентами a, b и c.
 */
 
-/** 1.Даны длины трёх сторон треугольника.
- * Определить, является ли треугольник прямоугольным.
- * @param name
- * @param callback_function
- */
 describe("is Triangle Right", () => {
   [
     [1, 1, 1, false, "not rectangular"],
@@ -37,15 +32,11 @@ describe("is Triangle Right", () => {
   });
 });
 
-/** 2.Пользователь вводит число R. Написать программу,
- * которая выведет в консоль длину окружности и площадь круга с радиусом R.
- */
 describe("calculate circumference and area of circle using prompted radius", () => {
-  const funcPropmpt = window.prompt;
-  let consoleSpy = jest.spyOn(console, "log");
+  let consoleSpy;
 
-  afterAll(() => {
-    window.prompt = funcPropmpt;
+  afterEach(() => {
+    consoleSpy.mockRestore();
   });
 
   [
@@ -54,8 +45,8 @@ describe("calculate circumference and area of circle using prompted radius", () 
     [10, 20 * Math.PI, 100 * Math.PI],
   ].forEach((item) => {
     it(`circle radius is ${item[0]} so circumference is ${item[1]} and area is ${item[2]}`, () => {
+      jest.spyOn(global.window, "prompt").mockImplementation(() => item[0]);
       consoleSpy = jest.spyOn(console, "log");
-      window.prompt = jest.fn(() => item[0]);
 
       printCircleSize();
 
@@ -66,15 +57,11 @@ describe("calculate circumference and area of circle using prompted radius", () 
   });
 });
 
-/** Пользователь вводит числа a, b и c.
- * Написать программу, выводящую корни квадратного уравнения с коэффициентами a, b и c.
- */
 describe("print roots of a quadratic equation with prompted coefficients", () => {
-  const funcPropmpt = window.prompt;
-  let consoleSpy = jest.spyOn(console, "log");
+  let consoleSpy;
 
-  afterAll(() => {
-    window.prompt = funcPropmpt;
+  afterEach(() => {
+    consoleSpy.mockRestore();
   });
 
   [
@@ -85,12 +72,11 @@ describe("print roots of a quadratic equation with prompted coefficients", () =>
     const [a, b, c, x1, x2, d] = item;
     it(`if a = ${a}, b = ${b}, c = ${c} then x1 = ${x1}, x2 = ${x2}, D = ${d}`, () => {
       consoleSpy = jest.spyOn(console, "log");
-      // window.prompt = jest.fn(() => item[0]);
-      window.prompt = jest
-        .fn()
-        .mockReturnValueOnce(a)
-        .mockReturnValueOnce(b)
-        .mockReturnValueOnce(c);
+      jest
+        .spyOn(global.window, "prompt")
+        .mockImplementationOnce(() => a)
+        .mockImplementationOnce(() => b)
+        .mockImplementationOnce(() => c);
 
       printRootsOfQEquation();
