--- src/hw-1/script2.js
@@ -1,3 +1,0 @@
-export default function printConcatLength(s1, s2) {
-  console.log(s1.length + s2.length);
-}
