--- src/hw-1/script3.js
@@ -1,3 +1,0 @@
-export default function printSummary() {
-  console.log("Here will be result");
-}
