--- src/index7.js
@@ -1,7 +1,0 @@
-import { addForm, addContent } from "./script7";
-// для браузера
-// import { addForm, addContent } from "./script7.js";
-
-addContent(document.querySelector("#content"));
-
-addForm(document.querySelector("#form"));
