--- src/index9.js
@@ -1,3 +1,0 @@
-import { printRootsOfQEquation } from "./script9";
-
-printRootsOfQEquation();
