--- src/script10.js
@@ -1,51 +1,0 @@
-export default function promptUserData() {
-  // date in posix
-  const rgexDate1 = /[0-9]{4}([-./])(0[1-9]|1[012])([-./])(0[1-9]|1[0-9]|2[0-9]|3[01])/;
-  const rgexDate2 = /(0?[1-9]|[12][0-9]|3[01])[-./](0?[1-9]|1[012])[-./]([0-9]{4})/;
-  const rgexDate3 = /(0?[1-9]|1[012])[-./](0?[1-9]|[12][0-9]|3[01])[-./]([0-9]{4})/;
-  /* сочинил сам (почти) */
-
-  const rgexEmail = /^((([0-9A-Za-z]{1}[-0-9A-z.]{1,}[0-9A-Za-z]{1})|([0-9А-Яа-я]{1}[-0-9А-я.]{1,}[0-9А-Яа-я]{1}))@([-A-Za-z]{1,}\.){1,2}[-A-Za-z]{2,})$/u;
-  /*  https://ru.stackoverflow.com/questions/571772/Регулярное-выражение-для-полной-проверки-email
-  разложим маску более детально
-  ^( - параметр что маска начинается с начала текста
-    (
-        (  - этот блок отвечает за логин латиницей
-            [0-9A-Za-z]{1} - 1й символ только цифра или буква
-            [-0-9A-z\.]{1,} - в середине минимум один символ (буква, цифра, _, -, .) (не менее 1 символа)
-            [0-9A-Za-z]{1} - последний символ только цифра или буква
-        )
-        | - параметр "или/или" выбирает блок "латиница" или "кирилица"
-        (  - этот блок отвечает за логин кирилицей
-            [0-9А-Яа-я]{1} - 1й символ только цифра или буква
-            [-0-9А-я\.]{1,} - в середине минимум один символ (буква, цифра, _, -, .) (не менее 1 символа)
-            [0-9А-Яа-я]{1} - последний символ только цифра или буква
-        )
-    )
-    @ - обазятельное наличие значка разделяющего логин от домена
-    (
-        [-0-9A-Za-z]{1,} - блок может состоять из "-", цифр и букв (не менее 1 символа)
-        \. - наличие точки в конце блока
-    ){1,2} - допускается от 1 до 2 блоков по вышеукащанной маске (mail. , ru.mail.)
-    [-A-Za-z]{2,} - блок описывайющий домен вехнего уровня (ru, com, net, aero etc) (не менее 2 символов)
-)$ - параметр что маска заканчивается в конце текста
-/u - параметр позволяющий работать с кирилицей
-*/
-
-  //  const rgexPhone = /(\s*\+\d)?^([\d\(])([\d\-\s_():=\+]{3,})/;
-  const rgexPhone = /^(\s*)([+(\d])([\d-\s_():=+]{3,})/;
-  /* в помощь https://ru.stackoverflow.com/questions/107021/Шаблон-для-номера-телефона */
-
-  const input = prompt("Enter your birthdate or email or phone:", "");
-
-  if (rgexDate1.test(input) || rgexDate2.test(input) || rgexDate3.test(input)) {
-    return "DATE";
-  }
-  if (rgexEmail.test(input)) {
-    return "EMAIL";
-  }
-  if (rgexPhone.test(input)) {
-    return "PHONE";
-  }
-  return "UNKNOWN";
-}
