--- src/script5.js
@@ -1,50 +1,0 @@
-/* Домашнее задание №5 «Массивы»
-Создайте массив целых чисел из 10 элементов.
-1.Выведите в консоль сумму всех элементов массива.
-2.Создайте новый массив на основе исходного, в
-котором каждый элемент будет вдвое больше
-элемента исходного массива с таким же индексом.
-(a[1] = 3, b[1] = 6, где a — исходный массив, b — новый
-массив).
-3.*Найдите и выведите в консоль наибольший и
-наименьший элементы исходного массива.
-*/
-
-/** Создайте массив целых чисел из 10 элементов.
- * 1.Выведите в консоль сумму всех элементов массива.
- */
-export function createArray() {
-  const arrTen = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9];
-  const summator = (result, thisValue) => result + thisValue;
-
-  console.log(arrTen.reduce(summator, 0));
-
-  return arrTen;
-}
-
-/** 2.Создайте новый массив на основе исходного, в котором каждый элемент будет
- * вдвое больше элемента исходного массива с таким же индексом.
- * (a[1] = 3, b[1] = 6, где a — исходный массив, b — новый массив).
- */
-export function createArrayByTemplate() {
-  return createArray().map((item) => 2 * item);
-}
-
-/** 3.*Найдите и выведите в консоль наибольший и
- * наименьший элементы исходного массива.
- */
-export function printExtremum() {
-  let minVal = 0;
-  let maxVal = 0;
-
-  createArray().forEach((item) => {
-    if (item < minVal) {
-      minVal = item;
-    }
-    if (item > maxVal) {
-      maxVal = item;
-    }
-  });
-
-  console.log(minVal, maxVal);
-}
