--- src/script6.js
@@ -1,36 +1,0 @@
-/* Домашнее задание №6 «Функции»
-1.Напишите функцию diff, которая получает в качестве
-параметров 2 числа и возвращает разницу между
-наибольшим и наименьшим.
-2.Напишите функцию isWord, которая принимает на
-вход текстовую строку. Функция возвращает true, если
-строка состоит из одного слова и false, если из
-нескольких.
-*Напишите функцию pow(a, x), которая вернёт
-значение числа a, возведённого в степень x.
-*/
-
-/** 1.Напишите функцию diff, которая получает в качестве
-параметров 2 числа и возвращает разницу между
-наибольшим и наименьшим.
-*/
-export function diff(a, b) {
-  return Math.abs(a - b);
-}
-
-/** 2.Напишите функцию isWord, которая принимает на
-вход текстовую строку. Функция возвращает true, если
-строка состоит из одного слова и false, если из
-нескольких.
-*/
-export function isWord(s) {
-  // console.log(s, s.split('\s').length);
-  return s.split("s").length === 1;
-}
-
-/** *Напишите функцию pow(a, x), которая вернёт
-значение числа a, возведённого в степень x.
-*/
-export function pow(a, x) {
-  return a ** x;
-}
